+++ Data/DAO/Interface/IPessoaTelefone.cs
@@ -4,7 +4,7 @@ namespace Cadastro_Teleatendimento.Data.DAO.Interface
 {
   public interface IPessoaTelefone<T>
   {
-    Pessoa? BuscaPorCpf(int cpf);
+    Pessoa? BuscaPorCpf(Int64 cpf);
     bool CadastraTelefones(int Fk_Pessoa, int Fk_Telefone);
     bool Insira(T item);
     bool Altere(T item);
